channel_migration/requirements.txt
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
mutagen==1.46.0
pycryptodomex==3.17
requests==2.28.2
urllib3==1.26.15
websockets==11.0.2

yt-dlp==2023.3.4

jsonlines~=3.1.0
scrapetube~=2.3.1
mysql-connector-python~=8.0.33